backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-multipart==0.0.9
pyyaml==6.0.2
pandas==2.2.2
